Add float suffixes and remove casts

+++ hdf/test/tvattr.c
@@ -67,7 +67,7 @@ char    data2[6] = {'A', 'B', 'C', 'D', 'E', 'F'}, idata2[6];
 uint16  attr2[2] = {16, 32}, iattr2[2];
 uint32  attr1[4] = {100, 132, 10032, 10064}, iattr1[2];
 char    attr3[6] = {'m', 'N', 'p', 'S', 't', '\0'}, iattr3[6];
-float32 attr4[2] = {(float32)32.001, (float32)-34.002}, iattr4[2];
+float32 attr4[2] = {32.001F, -34.002F}, iattr4[2];
 float64 attr5[2] = {64.12345, -64.12345}, iattr5[2];
 
 static int  create_vset_stuff(void);
